+++ services/cliente_s.go
@@ -76,55 +76,3 @@ func DeletarCliente(id int) bool {
 	}
 	return false
 }
-
-// --- Use Case: Solicitar Atendimento ---
-// O cliente agora "solicita" fornecendo os dados do atendimento, incluindo o ID desejado.
-func SolicitarAtendimento(a Atendimento) error { // implementado
-	if a.ID == 0 {
-		return errors.New("ID do atendimento não pode ser 0")
-	}
-	if _, existe := atendimentos[a.ID]; existe {
-		return fmt.Errorf("atendimento com ID %d já existe", a.ID)
-	}
-	if _, ok := LerCliente(a.ClienteID); !ok {
-		return fmt.Errorf("cliente com ID %d nao encontrado", a.ClienteID)
-	}
-	// Define o status padrão para uma solicitação
-	a.Status = "Solicitado"
-	atendimentos[a.ID] = a
-	//fmt.Printf("Atendimento (ID: %d) solicitado pelo cliente %d.\n", a.ID, a.ClienteID)
-	return nil
-}
-
-// --- Use Case: Visualizar Documentos ---
-func ListarDocumentosCliente(clienteID int) []Documento { // implementado
-	var docsCliente []Documento
-	for _, atendimento := range atendimentos {
-		if atendimento.ClienteID == clienteID {
-			for _, doc := range documentos {
-				if doc.AtendimentoID == atendimento.ID {
-					docsCliente = append(docsCliente, doc)
-				}
-			}
-		}
-	}
-	//fmt.Printf("Buscando documentos para o cliente %d...\n", clienteID)
-	return docsCliente
-}
-
-// MaiorIDCliente percorre o mapa de clientes e retorna o maior ID encontrado.
-func MaiorIDCliente() int {
-	// caso o mapa estar vazio.
-	if len(clientes) == 0 {
-		return 0
-	}
-
-	maiorID := 0
-	// itera apenas sobre as chaves do mapa 'clientes'.
-	for id := range clientes {
-		if id > maiorID {
-			maiorID = id
-		}
-	}
-	return maiorID
-}
